Move VLM Reranker implementation to dedicated module
